azure - refactoring
